Promedios solucionados compañeros 3<

+++ src/ui/Dashboard.java
@@ -18,9 +18,9 @@ public class Dashboard extends javax.swing.JFrame {
         @SuppressWarnings("unchecked")
         // <editor-fold defaultstate="collapsed" desc="Generated Code">
         private void initComponents() {
-            int cantPersonas = AltaPersonas.listaPersonas.size();
-            int cantVehiculos = AltaVehiculos.altaListaVehiculosAviones.size() + AltaVehiculos.altaListaVehiculosBarcos.size();
-            Double promedio;
+            double cantPersonas = AltaPersonas.listaPersonas.size();
+            double cantVehiculos = AltaVehiculos.altaListaVehiculosAviones.size() + AltaVehiculos.altaListaVehiculosBarcos.size();
+            double promedio;
             if (cantVehiculos == 0 || cantPersonas == 0){
                 promedio = 0.0;
             }else {
@@ -97,7 +97,7 @@ public class Dashboard extends javax.swing.JFrame {
 
             lblPersonasIngresadas.setFont(new java.awt.Font("Segoe UI Black", 0, 70)); // NOI18N
             lblPersonasIngresadas.setForeground(new java.awt.Color(255, 255, 255));
-            lblPersonasIngresadas.setText(Integer.toString(cantPersonas));
+            lblPersonasIngresadas.setText(Double.toString(cantPersonas));
 
             javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
             jPanel4.setLayout(jPanel4Layout);
@@ -166,7 +166,7 @@ public class Dashboard extends javax.swing.JFrame {
 
             lblVehiculosIngresados.setFont(new java.awt.Font("Segoe UI Black", 0, 70)); // NOI18N
             lblVehiculosIngresados.setForeground(new java.awt.Color(255, 255, 255));
-            lblVehiculosIngresados.setText(Integer.toString(cantVehiculos));
+            lblVehiculosIngresados.setText(Double.toString(cantVehiculos));
 
             javax.swing.GroupLayout jPanel7Layout = new javax.swing.GroupLayout(jPanel7);
             jPanel7.setLayout(jPanel7Layout);
